separate out sibling runs accordion into its own component, implement basic mobile design
